feat(lib1): loadFrame without javax dependency
There are now two overloads of `loadFrame()` one of which stays the same
functionally by using javax.ImageIO to read the image, while the other
simply expects an IntArray, where an Int is a pixel encoded in ARGB
order. This function can be called by the example Android app, as there
won't be the javax error.

Signed-off-by: Anton Kriese <anton.kriese@fu-berlin.de>