feat(lib1): use correct file handling in Main
also uses the new VideoGeneratorImpl.loadFrame() implementation that
takes ARGB pixel arrays.
Upon closing the app, there will be a file at
`/storage/emulated/0/Documents/exampleTestOutput/test.mkv` in the Device
Manager for the emulater which can be saved to the PCs filesystem and
looked at with some video software.

Signed-off-by: Anton Kriese <anton.kriese@fu-berlin.de>