test(lib2): Implement a class to generate random test cases
This class generates 2 videos that have some diffs/insertions/deletions and the correct alignment. The modifications are randomly inserted. Based on a state-machine-like machanism.
Also a test was implemented using this class. It would fail often because our algorithms does not do a perfect job in aligning the videos. And because sometimes the modifications of a frame have a larger difference to the original frame than another frame that might come later..
Thats why it does not assert and thus does not actually fail ;)
Signed-off-by: simonsasse <simonsasse97@gmail.com>