feat(gui): include lib2 in gui
This needed some adjustment of the other libraries' gradle settings, as the
subprojects' `settings.gradle.kts` are ignored when being used as a dependency.
That's why the VideoGenerator is defined as a dependency of the gui, even though
there is no direct connection between the gui and lib1. But because lib2 depends
on lib1 which is mentioned in lib2's `settings.gradle.kts` (which is not read by
the gui), the gui needs to define lib1's project path.

Signed-off-by: Anton Kriese <anton.kriese@fu-berlin.de>