doc(general): fix license checks
This is done by including certain packages in the
allowed licenses file. The plugin has trouble with
resolving license files for the mentioned packages,
which is mentioned in the new license README with a
link to the issue on GitHub.
I checked the mentioned packages (all jetbrains') and
they all offer the Apache 2.0 license, so this should
be fine.

Signed-off-by: Anton Kriese <anton.kriese@fu-berlin.de>