refactor(gui): make actual jump function private
The actual jump function (previously `jumpToFrame()`, now
`jump()`) should only be called from within the navigator.
Also, `jumpToFrame(index: Int)` can now be called from the
outside to jump to a specified frame. This change makes the
interface easier to work with, as the actual jump function
is now hidden.

Signed-off-by: Anton Kriese <anton.kriese@fu-berlin.de>