feat(lib2): signal singleton to algorithm
This introduces a singleton that can be used to signal
the premature cancellation of a running algorithm.
The algorithm checks if it is still alive regularly
while the user can signal from the outside that the
algorithm should finish itself.

This is similar to how one would cancel a coroutine
in compose (with `yield`), but we wanted to avoid having
kotlinx as a dependency for lib2.

Signed-off-by: Anton Kriese <anton.kriese@fu-berlin.de>