fix(gui): jump to the correct frame on OverviewBar click
This was a bug that came with the assumption that all
colored boxed in the OverviewBar get the same width in
pixels, which is not true. This issue becomes only clear
with long diff sequences as some boxes seem much bigger
than others (at most 1 pixel wider). We keep the boxes'
sizes, but save their offset in the row to determine which
box was actually clicked.
This is needed for the cursor drag feature as otherwise we
could simply give each box its own tapGesture event handler.

Signed-off-by: Anton Kriese <anton.kriese@fu-berlin.de>